+++ lib/pages/profile/formText.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/cupertino.dart';
 
 class FormText extends StatelessWidget {
   final double width;
